tests/wpt/web-platform-tests/referrer-policy/generic/subresource-test/image-decoding.html

<!DOCTYPE html>
<!-- TODO(kristijanburnik): Remove subres. duplication. Reuse a template. -->
<html>
  <head>
    <title>Image decoding - cross-origin image request</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/common/security-features/resources/common.js"></script>
    <script src="/referrer-policy/generic/referrer-policy-test-case.js"></script>
  </head>
  <body>
    <h1>Image decoding - cross-origin image request</h1>
    <p>If you can read JSON encoded HTTP headers of the image below,
       the decoding works as expected.</p>

    <pre id="received_message">Running...</pre>

    <script>
      promise_test(function() {
        var urlPath = '/common/security-features/subresource/image.py';
        var url = location.protocol + "//www1." + location.hostname + ":" + location.port +
                  urlPath + "?cache_destroyer=" + (new Date()).getTime();
        return requestViaImageForReferrerPolicy(url, undefined, "always")
          .then(function(message) {
            var pre = document.getElementById('received_message')
            var headers = message.headers;
            pre.innerHTML = "";
            pre.innerHTML += url + ":\n\n";
            pre.innerHTML += JSON.stringify(headers, null, 2) + "\n\n"
            assert_own_property(headers, "host")
            assert_own_property(headers, "connection")
          });
      }, "Image is encoding headers as JSON.");
    </script>

    <div id="log"></div>
  </body>
</html>
