third_party/WebKit/LayoutTests/platform/win/external/wpt/referrer-policy/generic/subresource-test/image-decoding-expected.txt

This is a testharness.js-based test.
PASS Image is encoding headers as JSON. 
Harness: the test ran to completion.
